+++ kivyblocks/download.py
@@ -1,0 +1,7 @@
+from kivy.properties import StringProperty
+
+class DownloadFile(ClickableIconText):
+	def __init__(self, **kw):
+		pass
+
+