+++ kivyblocks/ffpyplayer_video.py
@@ -6,12 +6,14 @@ from ffpyplayer.tools import set_log_callback
 
 from kivy.factory import Factory
 from kivy.app import App
+from kivy.core.window import Window
 from kivy.uix.image import Image
 from kivy.uix.widget import Widget
 from kivy.clock import Clock
 from kivy.properties import StringProperty, BooleanProperty, \
 			OptionProperty, NumericProperty
 from kivy.graphics.texture import Texture
+from kivy.graphics import Color, Line
 from kivyblocks.ready import WidgetReady
 
 
@@ -26,17 +28,17 @@ class FFVideo(WidgetReady, Image):
 	duration = NumericProperty(-1)
 	position = NumericProperty(-1)
 	volume = NumericProperty(-1)
-	framerate = NumericProperty(180)
 	in_center_focus = BooleanProperty(False)
 
 	def __init__(self, **kwargs):
 		self._player = None
 		self._update_task = None
-		self._texture = None
+		self.vh_task = None
 		self.is_black = False
 		self.videosize = None
-		self.timeperiod = 1.0 / self.framerate
-		self.ff_opts = {}
+		self.ff_opts = {
+			"framedrop":True
+		}
 		self.lib_opts = {}
 		self.headers_pattern = {}
 		super(FFVideo, self).__init__(**kwargs)
@@ -46,6 +48,7 @@ class FFVideo(WidgetReady, Image):
 		self.register_event_type('on_leave_focus')
 		self.register_event_type('on_enter_focus')
 		self.register_event_type('on_load_success')
+		self.register_event_type('on_startplay')
 
 	def on_open_failed(self, *args):
 		pass
@@ -101,10 +104,14 @@ class FFVideo(WidgetReady, Image):
 	def on_status(self, *args):
 		print('on_status called, ', self.status)
 		if self._player is None:
+			print('no _player, do nothing')
 			return
 
+		Window.allow_screensaver = True
 		if self.status == 'play':
+			Window.allow_screensaver = False
 			self._player.set_pause(False)
+			self.video_handle()
 		elif self.status == 'pause':
 			self._player.set_pause(True)
 		elif self.status == 'stop':
@@ -112,20 +119,35 @@ class FFVideo(WidgetReady, Image):
 		else:
 			pass
 
+	def on_parent(self, *args):
+		if self.parent:
+			return
+		self._play_stop()
+
+	def on_startplay(self, *args):
+		pass
+
 	def on_frame(self, *args):
 		if self._player is None:
 			return
-		if self.audio_id is None:
-			return
-		self._player.request_channel(self, 'audio', 'open', self.audio_id)
+		if not self.playing:
+			self.dispatch('on_startplay')
+			self._player.request_channel( \
+								'audio', 'open', self.audio_id)
+			self.seek(self.position)
+			self.playing = True
+		if self.duration > 0:
+			p = self.position / self.duration * self.width
+			self.canvas.after.clear()
+			with self.canvas.after:
+				Color(1,1,1,1)
+				Line()
+				Line(points=[0, 0, self.width, 0], width=1)
+				Color(1,0,0,1)
+				Line(points=[0,1,p,1], width=1)
 
 	def __del__(self):
-		if self._update_task:
-			self._update_task.cancel()
-			self._update_task = None
-		if self._player:
-			self._player.close_player()
-			self._player = None
+		self._play_stop()
 
 	def set_volume(self, v):
 		if self.play_mode == 'preview':
@@ -141,23 +163,27 @@ class FFVideo(WidgetReady, Image):
 		self.volume = v
 
 	def seek(self, pts):
+		self.vsync = False
 		if self.play_mode == 'preview':
 			return
 		if self._player is None:
 			return
 		if self.status != 'play':
 			return
-		self._player.seek(pts)
+		self._player.seek(pts, relative=False)
 		self.position = self._player.get_pts()
+		self.last_val = None
 
-	def mute(self, flag):
+	def mute(self, flag=None):
 		if self.play_mode == 'preview':
 			return
 		if self._player is None:
 			return
 		if self.status != 'play':
 			return
-		self._player.set_mute(flag)
+		x = self._player.get_mute()
+		print('Video(), mute=', x)
+		self._player.set_mute(not x)
 		
 	def switch_audio(self):
 		if self.play_mode == 'preview':
@@ -169,7 +195,9 @@ class FFVideo(WidgetReady, Image):
 		self._player.request_channel('audio', action='cycle')
 
 	def on_v_src(self, o, src):
-		self._play_stop()
+		self.status = 'stop'
+		self.playing = False
+
 		ff_opts = {
 			'pause':False
 		}
@@ -194,7 +222,8 @@ class FFVideo(WidgetReady, Image):
 		# self._player = MediaPlayer(self.v_src) 
 		self._player = MediaPlayer(self.v_src, ff_opts=ff_opts, \
 						lib_opts=lib_opts) 
-		self._play_start()
+		# self._play_start()
+		self.status = 'play'
 
 	def file_opened(self, files):
 		self.v_src = files[0]
@@ -202,7 +231,6 @@ class FFVideo(WidgetReady, Image):
 	def play(self):
 		if self._player is None:
 			return
-		# self._player.set_pause(False)
 		self.status = 'play'
 
 	def pause(self):
@@ -212,33 +240,32 @@ class FFVideo(WidgetReady, Image):
 		self.status = 'pause'
 
 	def _play_start(self):
-		self.timepass = 0.0
 		self.last_frame = None
 		self.is_black = False
-		self.first_play = True
-		self._update_task = Clock.schedule_interval(self._update, self.timeperiod)
+		self.vsync = False
 
 	def _get_video_info(self):
-		if self.first_play:
+		if not self.playing:
+			self.playing = True
 			meta = self._player.get_metadata()
 			self.duration = meta['duration']
-			self.position = 0
 			self._out_fmt = meta['src_pix_fmt']
 			self.frame_rate = meta['frame_rate']
 			self.videosize = meta['src_vid_size']
 
 	def _play_stop(self):
-		if self._player is None:
-			return
-		self._update_task.cancel()
-		self._update_task = None
-		self._player.close_player()
-		self._player = None
-		self.timepass = 0
+		if self._player:
+			self._player.close_player()
+			self._player = None
+		if self._update_task:
+			self._update_task.cancel()
+			self._update_task = None
+		if self.vh_task:
+			self.vh_task.cancel()
+			self.vh_task = None
 		self.next_frame = None
 		self.duration = -1
-		self.position = -1
-		self.frame_rate = None
+		self.position = 0
 		self.videosize = None
 		
 	def on_size(self, *args):
@@ -283,7 +310,6 @@ class FFVideo(WidgetReady, Image):
 		fbo['tex_y'] = 0
 		fbo['tex_u'] = 1
 		fbo['tex_v'] = 2
-		self._texture = fbo.texture
 		dy, du, dv, _ = img.to_memoryview()
 		if dy and du and dv:
 			self._tex_y.blit_buffer(dy, colorfmt='luminance')
@@ -304,35 +330,44 @@ class FFVideo(WidgetReady, Image):
 		self.texture = texture
 		# print('img_size=', w, h, 'window size=', self.size)
 
-	def _update(self, dt):
-		if self.last_frame is None:
-			frame, val = self._player.get_frame()
-			if val == 'eof':
-				print('*****EOF******')
-				self.status = 'stop'
-				self.set_black()
-				return
-			if val == 'pause':
-				self.status = 'pause'
-				return
-			if frame is None:
-				# print('video null', time.time())
-				self.set_black()
-				return
-			self.last_frame = frame
-			self.video_ts = val
-			self._get_video_info()
-			
-		self.timepass += self.timeperiod
+	def video_handle(self, *args):
+		if self._update_task:
+			self._update_task.cancel()
+			self._update_task = None
+		if self._player is None:
+			return 
+		frame, val = self._player.get_frame()
+		if val == 'eof':
+			self.status = 'stop'
+			self.set_black()
+			self.last_val = None
+			return
+		if val == 'pause':
+			self.status = 'pause'
+			self.last_val = None
+			return
+		if frame is None:
+			self.set_black()
+			self.last_val = None
+			self.vh_task = Clock.schedule_once(self.video_handle, 0.1)
+			return
+
+		self. _get_video_info()
+		self.last_frame = frame
+		self.video_ts = val
+		if self.last_val is None:
+			self.last_val = val
+			self._update_task = Clock.schedule_once(self.do_update, 0)
+		else:
+			t = val - self.last_val
+			if t > 0:
+				self._update_task = Clock.schedule_once(self.do_update, t)
+			else:
+				self._update_task = Clock.schedule_once(self.do_update, 0)
+
+	def do_update(self, *args):
 		self.position = self._player.get_pts()
-		if self.timepass < self.video_ts:
-			return
-
-		if self.timepass > self.video_ts +0.2:
-			self.last_frame = None
-			return
-
-		self.status = 'play'
+		self.volume = self._player.get_volume()
 		img, t = self.last_frame
 		if self._out_fmt == 'yuv420p':
 			self.show_yuv420(img)
@@ -340,5 +375,5 @@ class FFVideo(WidgetReady, Image):
 			self.show_others(img)
 		self.dispatch('on_frame', self.last_frame)
 		self.last_frame = None
-
-Factory.register('FFVideo', FFVideo)
+		self.vh_task = Clock.schedule_once(self.video_handle, 0)
+