+++ kivyblocks/hierarchy.py
@@ -13,31 +13,34 @@ from .scrollpanel import ScrollPanel
 from .clickable import SingleCheckBox
 from .baseWidget import Text
 from .utils import CSize
+from .command_action import cmd_action
 
 from appPublic.registerfunction import getRegisterFunctionByName
 
-class TreeViewComplexNode(BoxLayout, TreeViewLabel):
+class TreeViewComplexNode(BoxLayout, TreeViewNode):
 	otext = StringProperty(None)
+	font_size_c = NumericProperty(1)
+	node_height = NumericProperty(2)
 	checkbox = BooleanProperty(False)
 	icon = StringProperty(None)
 	data = DictProperty(None)
 	def __init__(self, **kw):
 		super(TreeViewComplexNode, self).__init__(**kw)
 		self.orientation = 'horizontal'
-		self.size_hint_x = None
 		if self.checkbox:
 			cb = SingleCheckBox(size_hint=(None,None))
 			cb.bind(on_press=self.set_checked)
 			self.add_widget(cb)
 		if self.icon:
 			img = AsyncImage(source=self.icon, size_hint=(None,None))
-			img.size = CSize(1,1)
+			img.size = CSize(self.font_size_c,self.font_size_c)
 			self.add_widget(img)
-		txt = Text(otext=self.otext, i18n=True)
-		txt.texture_update()
-		txt.size_hint = (None, None)
-		txt.size = txt.texture_size
+		txt = Text(otext=self.otext, i18n=True, wrap=True,
+				font_size=CSize(self.font_size_c),halign='left')
 		self.add_widget(txt)
+		self.size_hint_x = 1
+		self.size_hint_y = None
+		self.height = CSize(self.node_height)
 		
 	def set_checked(self, o):
 		if not self.data:
@@ -52,6 +55,8 @@ class Hierarchy(ScrollPanel):
 	params = DictProperty(None)
 	method = StringProperty('get')
 	idField = StringProperty(None)
+	node_height = NumericProperty(2)
+	font_size_c = NumericProperty(1)
 	textField = StringProperty(None)
 	data = ListProperty(None)
 	checkbox = BooleanProperty(False)
@@ -60,12 +65,16 @@ class Hierarchy(ScrollPanel):
 	def __init__(self, **kw):
 		self.register_event_type('on_press')
 		self.tree = TreeView(hide_root=True)
-		self.tree.size_hint = (None, None)
+		self.tree.size_hint = (1, None)
 		self.tree.bind(on_node_expand=self.check_load_subnodes)
 		self.tree.bind(selected_node=self.node_selected)
 		super(Hierarchy, self).__init__(inner=self.tree, **kw)
 		if self.url:
 			self.data = self.get_remote_data()
+
+	def on_size(self, *args):
+		self.tree.size_hint_x = 1
+		self.tree.width = self.width
 
 	def on_press(self, node):
 		print('selected node=', node)
@@ -95,12 +104,14 @@ class Hierarchy(ScrollPanel):
 	def create_new_node(self, data, node=None):
 		n = TreeViewComplexNode(otext=data[self.textField],
 			checkbox=self.checkbox,
+			node_height=self.node_height,
+			font_size_c=self.font_size_c,
 			icon=data.get('icon') or self.icon
 		)
 		n.data = data
-		n.width = self.tree.indent_start + \
-						self.tree.indent_level * n.level \
-						+ sum([i.width for i in n.children])
+		# n.width = self.tree.indent_start + \
+		#				self.tree.indent_level * n.level \
+		#				+ sum([i.width for i in n.children])
 		if node:
 			self.tree.add_node(n, node)
 		else:
@@ -148,50 +159,10 @@ class Menu(Hierarchy):
 	def on_press(self, node):
 		self.tree.deselect_node()
 		data = {}
-		dw = node.data.get('datawidget')
-		if dw:
-			data_widget = Factory.Blocks.getWidgetById(dw)
-			if data_widget:
-				vn = node.data.get('datamethod', 'getValue')
-				if hasattr(data_widget, vn):
-					f = getattr(data_widget, vn)
-					data = f()
-					if not isinstance(data, dict):
-						data = {}
-
-		url = node.data.get('url')
-		target = Factory.Blocks.getWidgetById(node.data.get('target',self.target),self)
-		if url:
-			params = node.data.get('params',{})
-			params.update(data)
-			blocks = Factory.Blocks()
-			desc = {
-				"widgettype":"urlwidget",
-				"options":{
-					"url":url,
-					"params":params
-				}
-			}
-			w = blocks.widgetBuild(desc)
-			if w and target:
-				target.add_widget(w)
-			return 
-
-		rfname = node.data.get('rfname')
-		if rfname:
-			f = getRegisterFunctionByName(rfname)
-			if f:
-				f(self, **data)
-			return
-		
-		script = node.data.get('script')
-		if script:
-			target_name = node.data.get('target', self.target)
-			target =  Factory.Blocks.getWidgetById(target_name, self)
-			data.update({'self':target})
-			if target:
-				eval(script,data)
-			return
+		data = node.data.copy()
+		if self.target and data.get('target') is None:
+			data['target'] = self.target
+		return cmd_action(data, self)
 		
 Factory.register('Hierarchy', Hierarchy)
 Factory.register('Menu', Menu)