+++ kivyblocks/toggleitems.py
@@ -1,6 +1,5 @@
 from functools import partial
 from kivy.logger import Logger
-from kivy.uix.behaviors import TouchRippleButtonBehavior
 from kivy.graphics import Color, Rectangle
 from kivy.uix.boxlayout import BoxLayout
 from kivy.factory import Factory
@@ -11,46 +10,7 @@ from kivyblocks.bgcolorbehavior import BGColorBehavior
 from kivyblocks.utils import CSize
 from kivyblocks.baseWidget import Box
 from kivyblocks.widget_css import WidgetCSS
-
-class PressableBox(TouchRippleButtonBehavior, Box):
-	normal_css = StringProperty("default")
-	actived_css = StringProperty("default")
-	box_actived = BooleanProperty(False)
-	def __init__(self,
-				border_width=1,
-				user_data=None,
-				radius=[],
-				**kw):
-		super(PressableBox, self).__init__(
-			padding=[border_width,
-			border_width,
-			border_width,
-			border_width],
-			radius=radius,
-			**kw)
-		self.border_width = border_width
-		self.user_data = user_data
-		self.actived = False
-		self.csscls = self.normal_css
-
-	def active(self, flag):
-		self.box_actived = flag
-
-	def on_box_actived(self, o, v):
-		if self.box_actived:
-			self.csscls = self.actived_css
-		else:
-			self.csscls = self.normal_css
-
-	def on_press(self,o=None):
-		self.box_actived = True
-
-
-	def setValue(self,d):
-		self.user_data = d
-
-	def getValue(self):
-		return self.user_data
+from kivyblocks.clickable import PressableBox
 
 """
 ToggleItems format: