+++ docs/changelog.md
@@ -5,3 +5,5 @@ add script to support local .tmpl and .dspy file translation. it can build dynam
 
 ## version 0.3.1
 * uses weakref to collect all the i18n widgets in i18n.py
+* show video play position
+